feat: add hooks
