+++ frontend/src/hooks/useSessionUser.js
@@ -1,0 +1,7 @@
+import { useSelector } from "react-redux";
+
+function useSessionUser() {
+  return useSelector((state) => state.session.user);
+}
+
+export default useSessionUser;
